Change to size_t

+++ src/circbuf.c
@@ -7,12 +7,12 @@
 static int circbuf_fill(circbuf* buf){
   CIRCBUF_CHECK(buf,-1);
 
-  uint pos = buf->read_pos;
-  uint read_size = buf->read_size;
+  size_t pos = buf->read_pos;
+  size_t read_size = buf->read_size;
   char* bytes = buf->bytes;
 
   CIRCBUF_CHECK(pos + read_size - 1 < buf->size,-1);
-  uint n_read = fread(bytes+pos,(size_t)1,read_size,buf->fd);
+  size_t n_read = fread(bytes+pos,(size_t)1,read_size,buf->fd);
   if(n_read < read_size) {
     buf->finished = 1;
     memset(bytes + pos + n_read,'\0',read_size - n_read);
@@ -21,17 +21,17 @@ static int circbuf_fill(circbuf* buf){
   return n_read;
 }
 
-static int circbuf_rewind(circbuf* buf, uint start){
+static int circbuf_rewind(circbuf* buf, size_t start){
   CIRCBUF_CHECK(buf,-1);
 
-  uint size = buf->size;
+  size_t size = buf->size;
   char* bytes = buf->bytes;
-  uint pos = buf->pos;
+  size_t pos = buf->pos;
 
   CIRCBUF_CHECK(start <= pos,-1);
   CIRCBUF_CHECK(start <= size,-1);
   
-  for(uint i=start;i<size;i++){
+  for(size_t i=start;i<size;i++){
     bytes[i-start] = bytes[i];
   }
 
@@ -40,7 +40,7 @@ static int circbuf_rewind(circbuf* buf, uint start){
   return 0;
 }
 
-circbuf* circbuf_create(char* bytes, uint size, uint read_size, FILE* fd){
+circbuf* circbuf_create(char* bytes, size_t size, size_t read_size, FILE* fd){
   CIRCBUF_CHECK(bytes,NULL);
   CIRCBUF_CHECK(size >= 4*read_size,NULL);
   CIRCBUF_CHECK(fd,NULL);
@@ -66,16 +66,16 @@ void circbuf_free(circbuf* buf){
 
 int circbuf_print(circbuf* c){
   CIRCBUF_CHECK(c,-1);
-  uint size = c->size;
+  size_t size = c->size;
   char* bytes = c->bytes;
-  uint pos = c->pos;
-  for(uint i=0;i<pos;i++){
+  size_t pos = c->pos;
+  for(size_t i=0;i<pos;i++){
     putchar(bytes[i]);
     putchar(' ');
   }
   putchar('.');
   putchar(' ');
-  for(uint i=pos;i<size;i++){
+  for(size_t i=pos;i<size;i++){
     putchar(bytes[i]);
     putchar(' ');
   }
@@ -88,19 +88,19 @@ char* circbuf_head(circbuf* buf){
   return buf->bytes + buf->pos;
 }
 
-char* circbuf_head_forward(circbuf* buf, uint n){
+char* circbuf_head_forward(circbuf* buf, size_t n){
   CIRCBUF_CHECK(buf,NULL);
   CIRCBUF_CHECK(n <= buf->read_size,NULL);
 
-  uint size = buf->size;
-  uint rest = size - buf->read_pos;
+  size_t size = buf->size;
+  size_t rest = size - buf->read_pos;
   
   if(rest < buf->read_size) {
     int rc = circbuf_rewind(buf,buf->pos + n - buf->read_size);
     CIRCBUF_CHECK(rc >= 0,NULL);
   }
 
-  uint pos = buf->pos;
+  size_t pos = buf->pos;
   pos += n;
 
   if(pos + buf->read_size - 1 >= buf->read_pos){
